Refactored KernelInputStream under stream package, refactored KernelMessageStream to KernelOutputStream
