+++ kernel/src/main/scala/com/ibm/spark/kernel/api/Kernel.scala
@@ -10,6 +10,7 @@ import com.ibm.spark.interpreter._
 import com.ibm.spark.kernel.protocol.v5
 import com.ibm.spark.kernel.protocol.v5.kernel.ActorLoader
 import com.ibm.spark.kernel.protocol.v5.magic.MagicParser
+import com.ibm.spark.kernel.protocol.v5.stream.KernelInputStream
 import com.ibm.spark.magic.{MagicLoader, MagicExecutor}
 import scala.language.dynamics
 
@@ -85,7 +86,7 @@ class Kernel (
       "The StreamInfo provided is not a KernelMessage instance!")
 
     val kernelMessage = streamInfo.asInstanceOf[v5.KernelMessage]
-    val outputStream = new v5.stream.KernelMessageStream(
+    val outputStream = new v5.stream.KernelOutputStream(
       actorLoader, v5.KMBuilder().withParent(kernelMessage),
       global.ScheduledTaskManager.instance, "stdout")
 
@@ -104,7 +105,7 @@ class Kernel (
       "The StreamInfo provided is not a KernelMessage instance!")
 
     val kernelMessage = streamInfo.asInstanceOf[v5.KernelMessage]
-    val outputStream = new v5.stream.KernelMessageStream(
+    val outputStream = new v5.stream.KernelOutputStream(
       actorLoader, v5.KMBuilder().withParent(kernelMessage),
       global.ScheduledTaskManager.instance, "stderr")
 
@@ -121,7 +122,7 @@ class Kernel (
     require(streamInfo.isInstanceOf[v5.KernelMessage],
       "The StreamInfo provided is not a KernelMessage instance!")
 
-    new v5.input.KernelInputStream(
+    new KernelInputStream(
       actorLoader,
       v5.KMBuilder()
         .withIds(streamInfo.asInstanceOf[v5.KernelMessage].ids)
